# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AdminLoginInput {
  /** @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$ */
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminIdentity {
  /** @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$ */
  email: string;
  fullName: string;
}

export type AdminSession = AdminIdentity & {
  token: string;
};

export interface MonthlyTrend {[key: string]: number}

export interface AdminDashboardStats {
  totalApplications: number;
  pendingReview: number;
  admitted: number;
  rejected: number;
  waitlisted: number;
  draft: number;
  monthlyTrend: MonthlyTrend;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  ADMITTED: 'ADMITTED',
  REJECTED: 'REJECTED',
  WAITLISTED: 'WAITLISTED',
} as const;

export interface Application {
  id: number;
  prn: string;
  firstName: string;
  lastName: string;
  otherNames: string;
  /** @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$ */
  email: string;
  phoneNumber: string;
  gender: string;
  dateOfBirth: string;
  nationality: string;
  district: string;
  subcounty: string;
  village: string;
  programChoice1: string;
  programChoice2: string;
  programChoice3: string;
  studyMode: string;
  academicYear: string;
  semester: string;
  emailVerified: boolean;
  status: ApplicationStatus;
  reviewStatus: string;
  /** @nullable */
  submittedAt: string | null;
  /** @nullable */
  reviewedAt: string | null;
  /** @nullable */
  reviewerNotes: string | null;
  uceResult: string;
  uaceResult: string;
  documents: string;
  extras: string;
  feePaid: number;
  feeRequired: number;
  feeCurrency: string;
  createdAt: string;
  updatedAt: string;
}

export interface PaginatedApplications {
  content: Application[];
  page: number;
  size: number;
  totalElements: number;
  totalPages: number;
}

export type ApplicationReviewInputReviewStatus = typeof ApplicationReviewInputReviewStatus[keyof typeof ApplicationReviewInputReviewStatus];


export const ApplicationReviewInputReviewStatus = {
  admitted: 'admitted',
  rejected: 'rejected',
  waitlisted: 'waitlisted',
} as const;

export interface ApplicationReviewInput {
  reviewStatus: ApplicationReviewInputReviewStatus;
  notes: string;
}

export type GetAdminApplicationsParams = {
status?: GetAdminApplicationsStatus;
search?: string;
/**
 * @minimum 0
 */
page?: number;
/**
 * @minimum 1
 * @maximum 50
 */
size?: number;
};

export type GetAdminApplicationsStatus = typeof GetAdminApplicationsStatus[keyof typeof GetAdminApplicationsStatus];


export const GetAdminApplicationsStatus = {
  ALL: 'ALL',
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  ADMITTED: 'ADMITTED',
  REJECTED: 'REJECTED',
  WAITLISTED: 'WAITLISTED',
} as const;

export type GetAdminRecentApplicationsParams = {
/**
 * @minimum 1
 * @maximum 10
 */
limit?: number;
};