# lib/api-zod/src/generated/types/adminDashboardStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyTrend } from './monthlyTrend';

export interface AdminDashboardStats {
  totalApplications: number;
  pendingReview: number;
  admitted: number;
  rejected: number;
  waitlisted: number;
  draft: number;
  monthlyTrend: MonthlyTrend;
}