# lib/api-zod/src/generated/types/application.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStatus } from './applicationStatus';

export interface Application {
  id: number;
  prn: string;
  firstName: string;
  lastName: string;
  otherNames: string;
  /** @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$ */
  email: string;
  phoneNumber: string;
  gender: string;
  dateOfBirth: string;
  nationality: string;
  district: string;
  subcounty: string;
  village: string;
  programChoice1: string;
  programChoice2: string;
  programChoice3: string;
  studyMode: string;
  academicYear: string;
  semester: string;
  emailVerified: boolean;
  status: ApplicationStatus;
  reviewStatus: string;
  /** @nullable */
  submittedAt: string | null;
  /** @nullable */
  reviewedAt: string | null;
  /** @nullable */
  reviewerNotes: string | null;
  uceResult: string;
  uaceResult: string;
  documents: string;
  extras: string;
  feePaid: number;
  feeRequired: number;
  feeCurrency: string;
  createdAt: string;
  updatedAt: string;
}