# lib/api-zod/src/generated/types/applicationReviewInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationReviewInputReviewStatus } from './applicationReviewInputReviewStatus';

export interface ApplicationReviewInput {
  reviewStatus: ApplicationReviewInputReviewStatus;
  notes: string;
}