# lib/api-zod/src/generated/types/paginatedApplications.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';

export interface PaginatedApplications {
  content: Application[];
  page: number;
  size: number;
  totalElements: number;
  totalPages: number;
}